refactor leaderboard utils
