+++ lib/leaderboardUtils.js
@@ -1,5 +1,4 @@
-
-const categories = [
+export const categories = [
   { slug: "eod_update", title: "EOD Updates" },
   { slug: "pr_opened", title: "Pull Requests Opened" },
   { slug: "pr_merged", title: "Pull Requests Merged" },
@@ -9,4 +8,13 @@ const categories = [
   { slug: "comment_created", title: "Comments Created" },
 ];
 
-export default categories;
+export const getMonthRepresentation = (date) => {
+  let d = new Date(date);
+  // Use previous month if before 14th of current month.
+  if (d.getDate() < 14) {
+    d = new Date(date.getFullYear(), date.getMonth() - 1, 1);
+  }
+  return d
+    .toLocaleDateString(undefined, { month: "long", year: "2-digit" })
+    .replace(" ", " ’");
+};
